langchain/requirements.txt
python-dotenv==1.0.1
langchain==0.3.0
langchain_community==0.3.0
langchain_openai==0.2.0
langchain-ollama==0.2.0
pinecone-client==2.2.1
openai==1.45.0